API/requirements.txt
flask-sock==0.6.0
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
pytz==2021.3
